+++ flutter/waitingnow/lib/src/app.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:waitingnow/src/screen/home.dart';
+import 'package:waitingnow/src/screen/waiting/home.dart';
 
 // stle 자동완성
 class MyApp extends StatelessWidget {
